requirements.txt
Flask==3.1.0
gunicorn
fastapi
uvicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
line-bot-sdk